reflect_tools: refactor to_string code
