+++ module/core/reflect_tools/src/reflect/to_string.rs
@@ -6,17 +6,9 @@
 pub( crate ) mod private
 {
 
-  #[ allow( unused_imports ) ]
-  use crate::
-  {
-    Ref,
-  };
-
   use std::
   {
     fmt,
-    // collections::HashMap,
-    // borrow::Cow,
   };
 
   // ==
@@ -32,122 +24,6 @@ pub( crate ) mod private
   /// Marker type for usign Well formatting.
   #[ derive( Debug, Default, Clone, Copy ) ]
   pub struct WithWell;
-
-  // ==
-
-  /// Marker type for ToStringWithFallback.
-  #[ derive( Debug, Default, Clone, Copy ) ]
-  pub struct ToStringWithFallbackParams< How, Fallback >( ::core::marker::PhantomData< fn() -> ( How, Fallback ) > );
-
-  /// Trait to convert a type to a string with a fallback formatting.
-  pub trait ToStringWithFallback< How, Fallback >
-  {
-    /// Converts the type to a string using the specified formatting or a fallback.
-    fn to_string_with_fallback( self ) -> String
-    ;
-  }
-
-  impl< T, How, Fallback > ToStringWithFallback< How, Fallback >
-  for Ref< '_, T, ToStringWithFallbackParams< How, Fallback > >
-  where
-    T : ToStringWith< Fallback >,
-  {
-    /// Converts the type to a string using the specified formatting.
-    fn to_string_with_fallback( self ) -> String
-    {
-      < T as ToStringWith< Fallback > >::to_string_with( self.0 )
-    }
-  }
-
-  impl< T, How, Fallback > ToStringWithFallback< How, Fallback >
-  for &Ref< '_, T, ToStringWithFallbackParams< How, Fallback > >
-  where
-    T : ToStringWith< How >,
-  {
-    /// Converts the type to a string using the fallback formatting.
-    fn to_string_with_fallback( self ) -> String
-    {
-      < T as ToStringWith< How > >::to_string_with( self.0 )
-    }
-  }
-
-  //
-
-  /// Macro to convert a value to a string using a specified formatting method with a fallback.
-  ///
-  /// # Parameters
-  /// - `$how`: The primary formatting type (e.g., `WithDebug`, `WithDisplay`).
-  /// - `$fallback`: The fallback formatting type.
-  /// - `$src`: The source value to format.
-  ///
-  /// # Example
-  /// ```rust
-  /// use core::fmt;
-  /// use reflect_tools::
-  /// {
-  ///   WithDebug,
-  ///   WithDisplay,
-  ///   to_string_with_fallback,
-  /// };
-  ///
-  /// // Define a struct that implements both Debug and Display traits.
-  /// struct Both;
-  ///
-  /// impl fmt::Debug for Both
-  /// {
-  ///   fn fmt( &self, f : &mut fmt::Formatter< '_ > ) -> fmt::Result
-  ///   {
-  ///     write!( f, "This is debug" )
-  ///   }
-  /// }
-  ///
-  /// impl fmt::Display for Both
-  /// {
-  ///   fn fmt( &self, f : &mut fmt::Formatter< '_ > ) -> fmt::Result
-  ///   {
-  ///     write!( f, "This is display" )
-  ///   }
-  /// }
-  ///
-  /// // Define a struct that implements only the Debug trait.
-  /// struct OnlyDebug;
-  ///
-  /// impl fmt::Debug for OnlyDebug
-  /// {
-  ///   fn fmt( &self, f : &mut fmt::Formatter< '_ > ) -> fmt::Result
-  ///   {
-  ///     write!( f, "This is debug" )
-  ///   }
-  /// }
-  ///
-  /// // Example usage: Using Both which implements both Debug and Display.
-  /// let src = Both;
-  /// let got = to_string_with_fallback!( WithDisplay, WithDebug, src );
-  /// let exp = "This is display".to_string();
-  /// // The primary formatting method WithDisplay is used.
-  /// assert_eq!( got, exp );
-  ///
-  /// // Example usage: Using OnlyDebug which implements only Debug.
-  /// let src = OnlyDebug;
-  /// let got = to_string_with_fallback!( WithDisplay, WithDebug, src );
-  /// let exp = "This is debug".to_string();
-  /// // The primary formatting method WithDisplay is not available, so the fallback WithDebug is used.
-  /// assert_eq!( got, exp );
-  /// ```
-
-  #[ macro_export( local_inner_macros ) ]
-  macro_rules! to_string_with_fallback
-  {
-    ( $how : ty, $fallback : ty, $src : expr )
-    =>
-    {{
-      use reflect_tools::ToStringWithFallback;
-      (
-        &reflect_tools::Ref::< '_, _, reflect_tools::ToStringWithFallbackParams< $how, $fallback > >::from( &$src )
-      )
-      .to_string_with_fallback()
-    }};
-  }
 
   // ==
 
@@ -212,13 +88,8 @@ pub mod exposed
     WithDebug,
     WithDisplay,
     WithWell,
-    ToStringWithFallbackParams,
-    ToStringWithFallback,
     ToStringWith,
   };
-  #[ doc( inline ) ]
-  #[ allow( unused_imports ) ]
-  use crate::to_string_with_fallback;
 }
 
 /// Prelude to use essentials: `use my_module::prelude::*`.
