+++ module/core/reflect_tools/src/reflect/to_string_with_fallback/params.rs
@@ -1,0 +1,7 @@
+//!
+//! Marker type for trait `_ToStringWithFallback` with type parameters.
+//!
+
+/// Marker type for trait `_ToStringWithFallback` with type parameters.
+#[ derive( Debug, Default, Clone, Copy ) ]
+pub struct ToStringWithFallbackParams< How, Fallback >( ::core::marker::PhantomData< fn() -> ( How, Fallback ) > );
